removed unnecessary static_cast

+++ src/str_to_short.cpp
@@ -6,7 +6,7 @@
 /*   By: jsaariko <jsaariko@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/03/31 13:39:29 by jsaariko      #+#    #+#                 */
-/*   Updated: 2021/03/31 13:48:16 by jsaariko      ########   odam.nl         */
+/*   Updated: 2021/03/31 13:50:14 by jsaariko      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,5 +22,5 @@ uint16_t str_to_short(std::string value) {
     if (num == 0 && value != "0") {
         throw std::exception();
     }
-    return (static_cast<uint16_t>(num));
+    return (num);
 }
